blog migrations updated

+++ modules/atlantis/blog/src/Module/Blog/Migrations/02_18_2016_130443_create_blog_config_table.php
@@ -14,20 +14,23 @@ class CreateBlogConfigTable extends Migration {
    * Run the migrations.
    */
   public function up() {
-
-    Schema::create('blog_config', function(\Illuminate\Database\Schema\Blueprint $table) {
-      $table->increments('id');
-      $table->string("config_key", 255);
-      $table->text("config_value")->nullable();
-      $table->timestamps();    
-    });
+    if (!Schema::hasTable('blog_config')) {
+      Schema::create('blog_config', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->increments('id');
+        $table->string("config_key", 255);
+        $table->text("config_value")->nullable();
+        $table->timestamps();
+      });
+    }
   }
 
   /**
    * Reverse the migrations.
    */
   public function down() {
-    Schema::drop('blog_config');
+    if (Schema::hasTable('blog_config')) {
+      Schema::drop('blog_config');
+    }
   }
 
 }
