+++ modules/atlantis/blog/src/Module/Blog/Migrations/2016_06_15_172722_add_gallery_id.php
@@ -4,27 +4,26 @@ use Illuminate\Database\Migrations\Migration;
 
 class AddGalleryId extends Migration {
 
-        /**
-        * Run the migrations.
-        */
-        public function up()
-        {
-                //
-                Schema::table('blog', function(\Illuminate\Database\Schema\Blueprint $table)
-                {                   
-                        $table->integer("gallery_id");
-                });
-        }
+  /**
+   * Run the migrations.
+   */
+  public function up() {
+    if (!Schema::hasColumn('blog', 'gallery_id')) {
+      Schema::table('blog', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->integer("gallery_id");
+      });
+    }
+  }
 
-        /**
-        * Reverse the migrations.
-        */
-        public function down()
-        {
-                 Schema::table('blog', function(\Illuminate\Database\Schema\Blueprint $table)
-                {                   
-                        $table->dropColumn("gallery_id");
-                });
-        }
+  /**
+   * Reverse the migrations.
+   */
+  public function down() {
+    if (Schema::hasColumn('blog', 'gallery_id')) {
+      Schema::table('blog', function(\Illuminate\Database\Schema\Blueprint $table) {
+        $table->dropColumn("gallery_id");
+      });
+    }
+  }
 
 }
